Removed use of CRAM "CORE" block while encoding.
Previously the cram_stats_encoding() function was written to
roughly minimise the storage requirements.  As such it used HUFFMAN or
BETA codec for data low-volume series and EXTERNAL for larger ones.
This avoided any block overheads.  (Actually there was a lot of
additional commented-out code to deal with subexponential codec too.)

However this comes at the price that (more or less) any field may
potentially be mixed with with any other field in the CORE block.
Checking a random file I see this is so:

$ cram_dump /tmp/_2.cram | egrep 'HUFFMAN|BETA'|less
         RL =>          HUFFMAN {1, 100, 1, 0}
         BA =>          HUFFMAN {4, 65, 67, 71, 84, 4, 1, 2, 3, 3}
         DL =>          HUFFMAN {4, 1, 2, 3, 4, 4, 1, 2, 3, 3}
         RG =>          HUFFMAN {1, 0, 1, 0}
         RI =>          HUFFMAN {1, 0, 1, 0}
         ...

Huffman 1,100, 1,0 is fine (1 symbol [100] with 1 bit-length [0];
so no storage in CORE), but DL and BA series here both require storage
and are interleaved into CORE.  Now BA and DL will be stored with type
EXTERNAL.

The change in file size is expected to be minimal; around 0.01% in
tests and cpu overhead is equally tiny.

The advantage is that is simplifies partial decoding.  Although the
cram_dependent_data_series() works out which data series are
co-locating in the same block, it becomes more efficient to do
partial decodes as fewer items share a common block.  It also becomes
much easier to filter by block, dropping specific data series
completely while retaining others.